5-RAG-ShopAssistant-GenAI/ShopAssistantChatbot2.0-Sndev/pyproject.toml
[project]
name = "customercarechatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "mysql-connector-python>=9.3.0",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.3",
]
